# Codechange: Don't scan vehicle pool to find targeting disaster vehicle when deleting any vehicle.