@rubidium42 2024-02-11
I know it's (very) unlikely with the current configurations, but have you considered what happens when there are multiple small UFO's targetting the same road vehicle?
Is it worth it preventing an already targeted vehicle to get targeted? And just removing any duplicates upon save load?