@LordAro 2024-02-11
How about changing disaster vehicles to better cope with invalid vehicles? They could retarget or just fly off the map