@PeterN 2024-02-11
For now I made it so a road vehicle which is already crashed or already a target is no longer targetted.

Changing how it behaves when it can't find a target is probably for another PR.